Moving useTheme out of index file, revert tsconfig
