+++ packages/components/src/components/Alert/Alert.tsx
@@ -9,10 +9,11 @@ import {
 } from 'react-native'
 import React, { FC, useState } from 'react'
 
-import { BaseColor, useColorScheme, useTheme } from '../../utils'
+import { BaseColor, useColorScheme } from '../../utils'
 import { Button, ButtonProps, ButtonVariants } from '../Button/Button'
 import { Icon, IconProps } from '../Icon/Icon'
 import { Spacer } from '../Spacer/Spacer'
+import { useTheme } from '../../utils/useTheme'
 
 /** Convenience function to set children content color correctly with light/dark mode */
 export const AlertContentColor = BaseColor
