+++ packages/components/src/components/SegmentedControl/SegmentedControl.tsx
@@ -4,7 +4,8 @@ import React, { FC, useEffect } from 'react'
 import styled from 'styled-components/native'
 
 import { ComponentWrapper } from '../../wrapper'
-import { PressableOpacityStyle, useTheme } from '../../utils'
+import { PressableOpacityStyle } from '../../utils'
+import { useTheme } from '../../utils/useTheme'
 
 /**
  * Props for {@link SegmentedControl}
