+++ packages/components/src/utils/useTheme.ts
@@ -1,0 +1,10 @@
+import { ColorSchemeName } from 'react-native'
+import { Theme, themes } from '@department-of-veterans-affairs/mobile-tokens'
+
+import { useColorScheme } from '.'
+
+/** Returns light/dark theme based on useColorScheme */
+export function useTheme(): Theme {
+  const themeName: ColorSchemeName = useColorScheme() || 'light'
+  return themes[themeName]
+}
